requirements.txt
Flask
openai
flask-wtf
Flask-Login
itsdangerous
concurrent.futures
requests
requests-oauthlib
firebase-admin
google-cloud-storage
google-auth
google-auth-oauthlib
werkzeug
authlib
logging
